# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/create": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Criar um novo usuário",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Simples Escolhas",
    "description": "Uma api para gerenciamento de pequenas eleições",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mario Silva",
            "description": "Nome completo do usuário"
          },
          "password": {
            "type": "string",
            "example": "senhaforte123",
            "description": "Senha (mínimo 8 caracteres)",
            "writeOnly": true
          },
          "login": {
            "type": "string",
            "example": "mario",
            "description": "Login único do usuário"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ],
            "example": "USER",
            "description": "Papel/role do usuário"
          }
        },
        "required": [
          "name",
          "password",
          "login",
          "role"
        ]
      }
    }
  }
}